fix(ui): fix img2img fit
