+++ invokeai/frontend/web/src/features/nodes/util/graphBuilders/buildTextToImageGraph.ts
@@ -2,16 +2,23 @@ import { RootState } from 'app/store/store';
 import {
   CompelInvocation,
   Graph,
+  IterateInvocation,
   LatentsToImageInvocation,
+  NoiseInvocation,
+  RandomIntInvocation,
+  RangeOfSizeInvocation,
   TextToLatentsInvocation,
 } from 'services/api';
 import { NonNullableGraph } from 'features/nodes/types/types';
-import { addNoiseNodes } from '../nodeBuilders/addNoiseNodes';
 
 const POSITIVE_CONDITIONING = 'positive_conditioning';
 const NEGATIVE_CONDITIONING = 'negative_conditioning';
 const TEXT_TO_LATENTS = 'text_to_latents';
 const LATENTS_TO_IMAGE = 'latents_to_image';
+const NOISE = 'noise';
+const RANDOM_INT = 'rand_int';
+const RANGE_OF_SIZE = 'range_of_size';
+const ITERATE = 'iterate';
 
 /**
  * Builds the Text to Image tab graph.
@@ -24,9 +31,14 @@ export const buildTextToImageGraph = (state: RootState): Graph => {
     cfgScale: cfg_scale,
     scheduler,
     steps,
+    width,
+    height,
+    iterations,
+    seed,
+    shouldRandomizeSeed,
   } = state.generation;
 
-  let graph: NonNullableGraph = {
+  const graph: NonNullableGraph = {
     nodes: {},
     edges: [],
   };
@@ -92,8 +104,203 @@ export const buildTextToImageGraph = (state: RootState): Graph => {
     },
   });
 
-  // Create and add the noise nodes
-  graph = addNoiseNodes(graph, TEXT_TO_LATENTS, state);
+  /**
+   * Now we need to handle iterations and random seeds. There are four possible scenarios:
+   * - Single iteration, explicit seed
+   * - Single iteration, random seed
+   * - Multiple iterations, explicit seed
+   * - Multiple iterations, random seed
+   *
+   * They all have different graphs and connections.
+   */
 
+  // Single iteration, explicit seed
+  if (!shouldRandomizeSeed && iterations === 1) {
+    // Noise node using the explicit seed
+    const noiseNode: NoiseInvocation = {
+      id: NOISE,
+      type: 'noise',
+      seed: seed,
+    };
+
+    graph.nodes[NOISE] = noiseNode;
+
+    // Connect noise to l2l
+    graph.edges.push({
+      source: { node_id: NOISE, field: 'noise' },
+      destination: {
+        node_id: TEXT_TO_LATENTS,
+        field: 'noise',
+      },
+    });
+  }
+
+  // Single iteration, random seed
+  if (shouldRandomizeSeed && iterations === 1) {
+    // Random int node to generate the seed
+    const randomIntNode: RandomIntInvocation = {
+      id: RANDOM_INT,
+      type: 'rand_int',
+    };
+
+    // Noise node without any seed
+    const noiseNode: NoiseInvocation = {
+      id: NOISE,
+      type: 'noise',
+    };
+
+    graph.nodes[RANDOM_INT] = randomIntNode;
+    graph.nodes[NOISE] = noiseNode;
+
+    // Connect random int to the seed of the noise node
+    graph.edges.push({
+      source: { node_id: RANDOM_INT, field: 'a' },
+      destination: {
+        node_id: NOISE,
+        field: 'seed',
+      },
+    });
+
+    // Connect noise to t2l
+    graph.edges.push({
+      source: { node_id: NOISE, field: 'noise' },
+      destination: {
+        node_id: TEXT_TO_LATENTS,
+        field: 'noise',
+      },
+    });
+  }
+
+  // Multiple iterations, explicit seed
+  if (!shouldRandomizeSeed && iterations > 1) {
+    // Range of size node to generate `iterations` count of seeds - range of size generates a collection
+    // of ints from `start` to `start + size`. The `start` is the seed, and the `size` is the number of
+    // iterations.
+    const rangeOfSizeNode: RangeOfSizeInvocation = {
+      id: RANGE_OF_SIZE,
+      type: 'range_of_size',
+      start: seed,
+      size: iterations,
+    };
+
+    // Iterate node to iterate over the seeds generated by the range of size node
+    const iterateNode: IterateInvocation = {
+      id: ITERATE,
+      type: 'iterate',
+    };
+
+    // Noise node without any seed
+    const noiseNode: NoiseInvocation = {
+      id: NOISE,
+      type: 'noise',
+    };
+
+    // Adding to the graph
+    graph.nodes[RANGE_OF_SIZE] = rangeOfSizeNode;
+    graph.nodes[ITERATE] = iterateNode;
+    graph.nodes[NOISE] = noiseNode;
+
+    // Connect range of size to iterate
+    graph.edges.push({
+      source: { node_id: RANGE_OF_SIZE, field: 'collection' },
+      destination: {
+        node_id: ITERATE,
+        field: 'collection',
+      },
+    });
+
+    // Connect iterate to noise
+    graph.edges.push({
+      source: {
+        node_id: ITERATE,
+        field: 'item',
+      },
+      destination: {
+        node_id: NOISE,
+        field: 'seed',
+      },
+    });
+
+    // Connect noise to t2l
+    graph.edges.push({
+      source: { node_id: NOISE, field: 'noise' },
+      destination: {
+        node_id: TEXT_TO_LATENTS,
+        field: 'noise',
+      },
+    });
+  }
+
+  // Multiple iterations, random seed
+  if (shouldRandomizeSeed && iterations > 1) {
+    // Random int node to generate the seed
+    const randomIntNode: RandomIntInvocation = {
+      id: RANDOM_INT,
+      type: 'rand_int',
+    };
+
+    // Range of size node to generate `iterations` count of seeds - range of size generates a collection
+    const rangeOfSizeNode: RangeOfSizeInvocation = {
+      id: RANGE_OF_SIZE,
+      type: 'range_of_size',
+      size: iterations,
+    };
+
+    // Iterate node to iterate over the seeds generated by the range of size node
+    const iterateNode: IterateInvocation = {
+      id: ITERATE,
+      type: 'iterate',
+    };
+
+    // Noise node without any seed
+    const noiseNode: NoiseInvocation = {
+      id: NOISE,
+      type: 'noise',
+      width,
+      height,
+    };
+
+    // Adding to the graph
+    graph.nodes[RANDOM_INT] = randomIntNode;
+    graph.nodes[RANGE_OF_SIZE] = rangeOfSizeNode;
+    graph.nodes[ITERATE] = iterateNode;
+    graph.nodes[NOISE] = noiseNode;
+
+    // Connect random int to the start of the range of size so the range starts on the random first seed
+    graph.edges.push({
+      source: { node_id: RANDOM_INT, field: 'a' },
+      destination: { node_id: RANGE_OF_SIZE, field: 'start' },
+    });
+
+    // Connect range of size to iterate
+    graph.edges.push({
+      source: { node_id: RANGE_OF_SIZE, field: 'collection' },
+      destination: {
+        node_id: ITERATE,
+        field: 'collection',
+      },
+    });
+
+    // Connect iterate to noise
+    graph.edges.push({
+      source: {
+        node_id: ITERATE,
+        field: 'item',
+      },
+      destination: {
+        node_id: NOISE,
+        field: 'seed',
+      },
+    });
+
+    // Connect noise to t2l
+    graph.edges.push({
+      source: { node_id: NOISE, field: 'noise' },
+      destination: {
+        node_id: TEXT_TO_LATENTS,
+        field: 'noise',
+      },
+    });
+  }
   return graph;
 };
